Move Observable into observables.py
