+++ janus_core/processing/correlator.py
@@ -181,9 +181,9 @@ class Correlation:
     ----------
     n_atoms : int
         Number of possible atoms to track.
-    a : tuple[Observable, dict]
+    a : Union[Observable, tuple[Observable, tuple, dict]]
         Getter for a and kwargs.
-    b : tuple[Observable, dict]
+    b : Union[Observable, tuple[Observable, tuple, dict]]
         Getter for b and kwargs.
     name : str
         Name of correlation.
@@ -199,6 +199,7 @@ class Correlation:
 
     def __init__(
         self,
+        *,
         n_atoms: int,
         a: Observable | tuple[Observable, tuple, dict],
         b: Observable | tuple[Observable, tuple, dict],
@@ -215,9 +216,9 @@ class Correlation:
         ----------
         n_atoms : int
             Number of possible atoms to track.
-        a : tuple[Observable, tuple, dict]
+        a : Union[Observable, tuple[Observable, tuple, dict]]
             Getter for a and kwargs.
-        b : tuple[Observable, tuple, dict]
+        b : Union[Observable, tuple[Observable, tuple, dict]]
             Getter for b and kwargs.
         name : str
             Name of correlation.
